a little fix

+++ cmd/main.go
@@ -11,7 +11,7 @@ func main() {
 	height := helper.ScanUint("Input the height of matrix: ")
 	width := helper.ScanUint("Input the width of matrix: ")
 
-	mtx := matrix.NewMatrix(uint(height), uint(width))
+	mtx := matrix.NewMatrix(height, width)
 
 	for _, v := range mtx.GetDiagonales() {
 		fmt.Println(v)
